requirements.txt
# Core dependencies
anthropic>=0.15.0
mcp>=0.1.0
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2

# Utils
python-dotenv>=1.0.0
aiohttp>=3.8.5
websockets>=11.0.3
starlette>=0.27.0

# Testing and development
pytest>=7.4.2
pytest-asyncio>=0.21.1
httpx>=0.25.0